Split Assert.Count.cs into focused files (#9220)
Refactors the 727-line Assert.Count.cs into three focused, per-concern files following the existing Assertions/ folder conventions:

- Assert.Count.InterpolatedStringHandlers.cs: the AssertCountInterpolatedStringHandler<TItem> struct.
- Assert.HasCount.cs: all HasCount public overloads, private HasCount core helpers, GetTrackedAssertionName, and ReportAssertCountFailed.
- Assert.IsEmpty.cs: IsEmpty and IsNotEmpty overloads, the AssertIsNotEmptyInterpolatedStringHandler<TItem> struct (kept with its only consumer), and ReportAssertIsNotEmptyFailed.

Assert.Count.cs is removed. Each new file is under 300 lines. No public API or behavior changes (purely a partial-class file reorganization).
